2024 11 22

+++ lib/myOdometer/src/myOdometry.cpp
@@ -90,11 +90,11 @@ void myOdometer::resetVelocity() {
 }
 
 float myOdometer::getLeftWheelVelocity() const {
-    return (float)leftWheelVelocity;
+    return leftWheelVelocity;
 }
 
 float myOdometer::getRightWheelVelocity() const {
-    return (float)rightWheelVelocity;
+    return rightWheelVelocity;
 }
